+++ src/ska_dlm/dlm_storage/dlm_storage_requests.py
@@ -1,7 +1,10 @@
 """DLM Storage API module."""
 
+import json
 import logging
+import os
 import random
+from contextlib import asynccontextmanager
 
 import requests
 from fastapi import FastAPI, Request
@@ -34,6 +37,28 @@ from ..exceptions import (
 logger = logging.getLogger(__name__)
 
 
+@asynccontextmanager
+async def lifespan(_: FastAPI):
+    """Startup function."""
+    location_config_path = os.getenv("LOCATION_CONFIG", None)
+    if location_config_path:
+        with open(location_config_path, "r", encoding="utf-8") as json_file:
+            location_json_config = json.loads(json_file.read())
+
+        for location in location_json_config:
+            _setup_location(location)
+
+    storage_config_path = os.getenv("STORAGE_CONFIG", None)
+    if storage_config_path:
+        with open(storage_config_path, "r", encoding="utf-8") as json_file:
+            storage_json_config = json.loads(json_file.read())
+
+        for storage in storage_json_config:
+            _setup_storage(storage)
+
+    yield  # Application runs here
+
+
 cli = ExceptionHandlingTyper()
 rest = fastapi_auto_annotate(
     FastAPI(
@@ -41,6 +66,7 @@ rest = fastapi_auto_annotate(
         description="REST interface of the SKA-DLM Storage Manager",
         version=ska_dlm.__version__,
         license_info={"name": "BSD-3-Clause", "identifier": "BSD-3-Clause"},
+        lifespan=lifespan,
     )
 )
 
@@ -91,6 +117,106 @@ def query_location_facility() -> list[str]:
     params = {"select": "id"}
     rows = DB.select("location_facility", params=params)
     return [row["id"] for row in rows]
+
+
+def _setup_location(location: dict):
+    """# noqa
+    Setup a location.
+
+    Parameters
+    ----------
+    location:
+        json object containing location configuration.
+            Example:
+                [{
+                    "name": "test_location",
+                    "type": "low-integration",
+                    "country": "AU",
+                    "city": "Perth",
+                    "facility": "local"
+                },{
+                    "name": "test_location2",
+                    "type": "local-dev",
+                    "country": "UK",
+                    "city": "Manchester",
+                    "facility": "local"
+                }]
+    """
+    location_query = query_location(location_name=location["name"])
+    if not location_query:
+        init_location(
+            location_name=location["name"],
+            location_type=location["type"],
+            location_country=location["country"],
+            location_city=location["city"],
+            location_facility=location["facility"],
+        )
+
+        logging.info("Location %s created.", {location["name"]})
+
+
+def _setup_storage(storage: dict):
+    """# noqa
+    Setup a storage endpoint.
+
+    Parameters
+    ----------
+    storage:
+        json object containing storage configuration.
+            Example:
+                [{
+                    "name": "test_source",
+                    "location": "test_location",
+                    "storage_type": "filesystem",
+                    "interface": "posix",
+                    "root_directory": "/",
+                    "config": {
+                        "name": "test_source",
+                        "type": "alias",
+                        "parameters":{"remote":"/"}
+                    }
+                },{
+                    "name": "s3_destination",
+                    "location": "test_location",
+                    "storage_type": "objectstore",
+                    "interface": "s3",
+                    "root_directory": "/",
+                        "config": {
+                        "name": "s3",
+                        "type": "s3",
+                        "parameters":{"access_key_id": "",
+                                      "provider": "Ceph",
+                                      "secret_access_key": "",
+                                      "endpoint": ""}
+                    }
+                }]
+
+    Raises:
+        ValueError: If the storage location is not defined.
+    """
+    store = query_storage(storage["name"])
+    if not store:
+        location_query = query_location(location_name=storage["location"])
+        if not location_query:
+            raise ValueError(f"Location {storage['location']} not defined")
+
+        store = init_storage(
+            storage_name=storage["name"],
+            location_id=location_query[0]["location_id"],
+            storage_interface=storage["interface"],
+            root_directory=storage["root_directory"],
+            storage_type=storage["storage_type"],
+        )
+        logging.info("Storage %s created.", {storage["name"]})
+    else:
+        store = store[0]["storage_id"]
+
+    store_config = get_storage_config(storage_id=store)
+    if not store_config:
+        create_storage_config(storage_id=store, config=storage["config"])
+        logging.info("Storage configuration for %s created.", {storage["name"]})
+
+    create_rclone_config(config=storage["config"])
 
 
 @cli.command()
@@ -225,6 +351,27 @@ def init_storage(
         ) from exc
 
     return DB.insert(CONFIG.DLM.storage_table, json=post_data)[0]["storage_id"]
+
+
+def set_storage_availability(storage_name: str, available: bool):
+    """
+    Set if storage is available.
+
+    Parameters
+    ----------
+    storage_name
+        name of the storage.
+    available
+        set storage availability.
+
+    Returns
+    -------
+    dict
+        the updated data item entry
+    """
+    params = {"storage_name": f"eq.{storage_name}"}
+    json_data = {"storage_available": available}
+    return DB.update(CONFIG.DLM.storage_table, params=params, json=json_data)[0]
 
 
 @cli.command()